test(radio): extract domain sorting tests to types.test.ts and add empty/badge queue tests
